pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.1"
description = "smaug tests"
authors = ["Chris Guida <chrisguida@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.2,<4.0"
black = "^24.3"
flake8 = "^7.0.0"
isort = "^5.13.2"
pre-commit = "^3.6.0"
pyln-testing = "^25.12"
pyln-client = "^25.12"
pyln-proto = "^25.12"
pytest = ">=7.4,<9"
pytest-timeout = "^2.1.0"
pytest-xdist = "^3.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
